Refactor srs get
